+++ shapeinference/dotgeneral.go
@@ -49,7 +49,7 @@ func DotGeneral(
 		lDim := lhs.Dimensions[axis]
 		rDim := rhs.Dimensions[rAxis]
 		if lDim == shapes.DynamicDim && rDim == shapes.DynamicDim {
-			if lhs.AxisName(axis) != rhs.AxisName(rAxis) {
+			if !shapes.AxisNameEqual(lhs.AxisName(axis), rhs.AxisName(rAxis)) {
 				return output, errors.Errorf("DotGeneral contracting axis #%d has different axis names (axis name conflict) for lhs (%q) and rhs (%q)", i, lhs.AxisName(axis), rhs.AxisName(rAxis))
 			}
 		} else if lDim != shapes.DynamicDim && rDim != shapes.DynamicDim && lDim != rDim {
@@ -73,13 +73,14 @@ func DotGeneral(
 		lDim := lhs.Dimensions[axis]
 		rDim := rhs.Dimensions[rAxis]
 		if lDim == shapes.DynamicDim && rDim == shapes.DynamicDim {
-			if lhs.AxisName(axis) != rhs.AxisName(rAxis) {
+			if !shapes.AxisNameEqual(lhs.AxisName(axis), rhs.AxisName(rAxis)) {
 				return output, errors.Errorf("DotGeneral batch axis #%d has different axis names (axis name conflict) for lhs (%q) and rhs (%q)", i, lhs.AxisName(axis), rhs.AxisName(rAxis))
 			}
 		} else if lDim != shapes.DynamicDim && rDim != shapes.DynamicDim && lDim != rDim {
 			return output, errors.Errorf("DotGeneral batch dimensions do not match: lhs[%d]=%d, rhs[%d]=%d", axis, lDim, rAxis, rDim)
 		}
 	}
+
 
 	isRHSUsed := make([]bool, rhs.Rank())
 	for _, axis := range rhsContractingAxes {